fix(playground): update saved model config to overwrite all model config if present
