+++ app/src/store/playground/index.ts
@@ -1,2 +1,3 @@
 export * from "./playgroundStore";
 export * from "./types";
+export * from "./playgroundStoreUtils";
